Removed out-of-scope tests

+++ testfiles/08_type_conversions/10_int_double_char_mix.c
@@ -2,5 +2,5 @@ int main() {
     int a = 2;
     double d = 2.0;
     char c = 1;
-    return (int)(d + a + c); // 5
+    return (d + a + c); // 5
 }
